ml-service/requirements.txt
# YOLO Face Detection Service Dependencies
# Python 3.10+ required

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# ML/Computer Vision
ultralytics==8.1.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless==4.9.0.80
numpy>=1.24.0
Pillow>=10.0.0

# Utilities
pydantic>=2.0.0
python-dotenv==1.0.0